refactor: 센서 값 타입 BigDecimal로 변경 및 예외 처리 테스트 추가
센서 값의 정밀도를 높이기 위해 `LatestSensorValueResponse`의 `value` 필드 타입을 `Double`에서 `BigDecimal`으로 변경합니다. 이 변경에 따라 `InfluxServiceImpl` 및 관련 테스트 코드(`InfluxLatestQueryTest`)를 수정하여 `BigDecimal` 타입을 처리하도록 했습니다.

또한, `SensorTypeControllerTest`에 서비스 계층에서 예외가 발생했을 때 HTTP 500 상태 코드를 반환하는지에 대한 테스트 케이스를 추가하여 안정성을 강화했습니다. 불필요한 주석 정리도 함께 진행했습니다.

+++ src/test/java/site/yesaido/cultivation_server/sensor/service/influx/InfluxLatestQueryTest.java
@@ -10,6 +10,7 @@ import site.yesaido.cultivation_server.sensor.dto.response.influx.LatestSensorVa
 import site.yesaido.cultivation_server.sensor.mapper.SensorValuePointMapper;
 import site.yesaido.cultivation_server.sensor.service.impl.InfluxServiceImpl;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 import java.util.Map;
@@ -51,7 +52,7 @@ class InfluxLatestQueryTest {
         List<LatestSensorValueResponse> result = service.findLatestByCultivationId(42L);
 
         assertThat(result).containsExactly(new LatestSensorValueResponse(
-                42L, "TEMPERATURE", "°C", 23.5,
+                42L, "TEMPERATURE", "°C", BigDecimal.valueOf(23.5),
                 Instant.parse("2026-08-09T12:34:56Z"),
                 "eui-01", "model-x", "sensor-01", "room-1", "farm-a"
         ));
